[7주차 과제 - todo] - post, comment 인피니티 스크롤 api 완료 - N + 1 problem 해결 - 레거시 코드 삭제
[work]
- api 테스트 및 검증

+++ spring-rest-api/src/main/java/org/restapi/springrestapi/service/comment/CommentService.java
@@ -7,7 +7,7 @@ import org.restapi.springrestapi.dto.comment.RegisterCommentRequest;
 
 public interface CommentService {
 	CommentResult registerComment(Long userId, RegisterCommentRequest request, Long postId);
-	CommentListResult getCommentList(Long postId, int cursor, int limit);
+	CommentListResult getCommentList(Long postId, Long cursor, int limit);
 	CommentResult updateComment(Long userId, PatchCommentRequest request, Long postId, Long id);
 	void deleteComment(Long userId, Long id, Long postId);
 }
